Refactor watcher log processor
