+++ cosmos/operators/_watcher/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+__all__ = ["get_xcom_val", "safe_xcom_push", "build_producer_state_fetcher"]
+
+from cosmos.operators._watcher.state import build_producer_state_fetcher, get_xcom_val, safe_xcom_push
